+++ codex/librarian/covers/purge.py
@@ -1,11 +1,10 @@
 """Purge comic covers."""
-from logging import getLogger
-
 from codex.librarian.covers import COVER_ROOT
 from codex.models import Comic
+from codex.settings.logging import get_logger
 
 
-LOG = getLogger(__name__)
+LOG = get_logger(__name__)
 
 
 def _cleanup_cover_dirs(path):
@@ -40,7 +39,7 @@ def purge_cover_paths(cover_paths):
 
 def purge_library_covers(library_pks):
     """Remove all cover thumbs for a library."""
-    LOG.verbose(f"Removing comic covers from libraries: {library_pks}")  # type: ignore
+    LOG.verbose(f"Removing comic covers from libraries: {library_pks}")
     cover_paths = Comic.objects.filter(library_id__in=library_pks).values_list(
         "cover_path", flat=True
     )