+++ codex/librarian/janitor/vacuum.py
@@ -1,13 +1,12 @@
 """Vacuum the database."""
-from logging import getLogger
-
 from django.db import connection
 from humanize import naturalsize
 
+from codex.settings.logging import get_logger
 from codex.settings.settings import BACKUP_DB_PATH, DB_PATH
 
 
-LOG = getLogger(__name__)
+LOG = get_logger(__name__)
 
 
 def vacuum_db():
@@ -19,7 +18,7 @@ def vacuum_db():
     new_size = DB_PATH.stat().st_size
     saved = naturalsize(old_size - new_size)
 
-    LOG.verbose(f"Vacuumed database. Saved {saved}.")  # type: ignore
+    LOG.verbose(f"Vacuumed database. Saved {saved}.")
 
 
 def backup_db():
@@ -27,4 +26,4 @@ def backup_db():
     BACKUP_DB_PATH.unlink(missing_ok=True)
     with connection.cursor() as cursor:
         cursor.execute(f"VACUUM INTO '{BACKUP_DB_PATH}'")
-    LOG.verbose("Backed up database.")  # type: ignore
+    LOG.verbose("Backed up database.")