+++ codex/librarian/searchd.py
@@ -2,7 +2,6 @@
 import os
 
 from datetime import datetime
-from logging import getLogger
 
 from django.core.management import call_command
 from django.utils import timezone
@@ -13,57 +12,68 @@ from codex.librarian.queue_mp import (
     SearchIndexRebuildIfDBChangedTask,
     SearchIndexUpdateTask,
 )
-from codex.models import LatestVersion, SearchResult
+from codex.models import LatestVersion, Library, SearchResult
+from codex.settings.logging import get_logger
 from codex.settings.settings import CACHE_PATH, XAPIAN_INDEX_PATH
 from codex.threads import QueuedThread
 
 
 SEARCH_INDEX_TIMESTAMP_PATH = CACHE_PATH / "search_index_timestamp"
 UPDATE_INDEX_DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S%Z"
-LOG = getLogger(__name__)
+LOG = get_logger(__name__)
 WORKERS = os.cpu_count()
 VERBOSITY = 1
 
 
 def update_search_index(rebuild=False):
     """Update the search index."""
-    XAPIAN_INDEX_PATH.mkdir(parents=True, exist_ok=True)
+    try:
+        any_update_in_progress = Library.objects.filter(
+            update_in_progress=True
+        ).exists()
+        if any_update_in_progress:
+            LOG.verbose("Database update in progress, not updating search index yet.")
+            return
 
-    if rebuild:
-        LOG.verbose("Rebuilding search index...")  # type: ignore
-        call_command(
-            "rebuild_index",
-            interactive=False,
-            workers=WORKERS,
-            verbosity=VERBOSITY,
-        )
-        LatestVersion.set_xapian_index_version()
-    else:
-        try:
-            timestamp = SEARCH_INDEX_TIMESTAMP_PATH.stat().st_mtime
-        except FileNotFoundError:
-            timestamp = 0
-        start = datetime.fromtimestamp(timestamp, tz=timezone.utc).strftime(
-            UPDATE_INDEX_DATETIME_FORMAT
-        )
+        XAPIAN_INDEX_PATH.mkdir(parents=True, exist_ok=True)
 
-        LOG.verbose(f"Updating search index since {start}...")  # type: ignore
-        # Workers are only possible with fork()
-        # django-haystack has a bug
-        # https://github.com/django-haystack/django-haystack/issues/1650
-        call_command(
-            "update_index",
-            "codex",
-            remove=True,
-            start=start,
-            workers=WORKERS,
-            verbosity=VERBOSITY,
-        )
-    SEARCH_INDEX_TIMESTAMP_PATH.touch()
+        if rebuild:
+            LOG.verbose("Rebuilding search index...")
+            call_command(
+                "rebuild_index",
+                interactive=False,
+                workers=WORKERS,
+                verbosity=VERBOSITY,
+            )
+            LatestVersion.set_xapian_index_version()
+        else:
+            try:
+                timestamp = SEARCH_INDEX_TIMESTAMP_PATH.stat().st_mtime
+            except FileNotFoundError:
+                timestamp = 0
+            start = datetime.fromtimestamp(timestamp, tz=timezone.utc).strftime(
+                UPDATE_INDEX_DATETIME_FORMAT
+            )
 
-    # Nuke the Search Result table as it's now out of date.
-    SearchResult.truncate_and_reset()
-    LOG.verbose("Updated search index.")  # type: ignore
+            LOG.verbose(f"Updating search index since {start}...")
+            # Workers are only possible with fork()
+            # django-haystack has a bug
+            # https://github.com/django-haystack/django-haystack/issues/1650
+            call_command(
+                "update_index",
+                "codex",
+                remove=True,
+                start=start,
+                workers=WORKERS,
+                verbosity=VERBOSITY,
+            )
+        SEARCH_INDEX_TIMESTAMP_PATH.touch()
+
+        # Nuke the Search Result table as it's now out of date.
+        SearchResult.truncate_and_reset()
+        LOG.verbose("Updated search index.")
+    except Exception as exc:
+        LOG.error(f"Update search index: {exc}")
 
 
 def rebuild_search_index_if_db_changed():
@@ -73,7 +83,7 @@ def rebuild_search_index_if_db_changed():
         task = SearchIndexUpdateTask(True)
         LIBRARIAN_QUEUE.put(task)
     else:
-        LOG.verbose("Database matches search index.")  # type: ignore
+        LOG.verbose("Database matches search index.")
 
 
 class SearchIndexer(QueuedThread):
@@ -81,7 +91,7 @@ class SearchIndexer(QueuedThread):
 
     NAME = "SearchIndexer"
 
-    def _process_item(self, task):
+    def process_item(self, task):
         """Run the updater."""
         if isinstance(task, SearchIndexRebuildIfDBChangedTask):
             rebuild_search_index_if_db_changed()
@@ -90,7 +100,7 @@ class SearchIndexer(QueuedThread):
         else:
             LOG.warning(f"Bad task sent to search index thread: {task}")
 
-    def run(self):
+    def run(self):  # TODO maybe redundant now
         """Run the multiprocessing start method change for haystack update_index."""
         force_darwin_multiprocessing_fork()
         super().run()